first cut at improving installer
The installer now takes ca-cert dirs and/or no-check-certificate so we
can have the option of a secure or insecure d/l.  The installer also
checks to see if the python msi or exe has been d/l before attempting to
download it.

Added support for the new exe installer in python 3.5